pyproject.toml
[project]
name = "template-langgraph-py"
version = "0.1.5"
description = ""
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asgi-correlation-id>=4.3.4",
    "blaxel[langchain,crewai]==0.1.6rc29",
    "copilotkit>=0.1.44",
    "crewai==0.102.0",
    "fastapi[standard]>=0.115.12",
    "langgraph-supervisor>=0.0.4",
    "nest-asyncio>=1.6.0",
    "rich>=13.9.4",
]
[dependency-groups]
dev = ["ruff>=0.8.2"]

[tool.ruff]
indent-width = 4

[tool.ruff.format]
indent-style = "space"
quote-style = "double"
skip-magic-trailing-comma = false

[tool.ruff.lint]
select = ["E", "F"]

[tool.hatch.metadata]
allow-direct-references = true
